Moves io.py into dlg.data
